// frontend/src/api/generated/model/importResult.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Family Finance Tracker
 * OpenAPI spec version: 0.1.0
 */
import type { ImportResultImportBatchId } from './importResultImportBatchId';
import type { UnmappedValuesOut } from './unmappedValuesOut';

export interface ImportResult {
  account_id: number;
  import_batch_id: ImportResultImportBatchId;
  total_rows_read: number;
  inserted: number;
  duplicates_skipped: number;
  unmapped: UnmappedValuesOut;
  errors: string[];
}